+++ iextrading4j-acceptance/src/test/java/pl/zankowski/iextrading4j/test/acceptance/v1/AlternativeAcceptanceTest.java
@@ -2,8 +2,12 @@ package pl.zankowski.iextrading4j.test.acceptance.v1;
 
 import org.junit.Test;
 import pl.zankowski.iextrading4j.api.alternative.CeoCompensation;
+import pl.zankowski.iextrading4j.api.alternative.CryptoBook;
+import pl.zankowski.iextrading4j.api.alternative.CryptoPrice;
 import pl.zankowski.iextrading4j.api.stocks.Quote;
 import pl.zankowski.iextrading4j.client.rest.request.alternative.CeoCompensationRequestBuilder;
+import pl.zankowski.iextrading4j.client.rest.request.alternative.CryptoBookRequestBuilder;
+import pl.zankowski.iextrading4j.client.rest.request.alternative.CryptoPriceRequestBuilder;
 import pl.zankowski.iextrading4j.client.rest.request.alternative.CryptoRequestBuilder;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -24,6 +28,22 @@ public class AlternativeAcceptanceTest extends IEXCloudV1AcceptanceTestBase {
                 .withSymbol("BTCUSDT")
                 .build());
         assertThat(result).isNotNull();
+    }
+
+    @Test
+    public void cryptoPriceTest() {
+        final CryptoPrice price = cloudClient.executeRequest(new CryptoPriceRequestBuilder()
+                .withSymbol("BTCUSD")
+                .build());
+        assertThat(price).isNotNull();
+    }
+
+    @Test
+    public void cryptoBookTest() {
+        final CryptoBook book = cloudClient.executeRequest(new CryptoBookRequestBuilder()
+                .withSymbol("BTCUSD")
+                .build());
+        assertThat(book).isNotNull();
     }
 
 }