+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/mapper/CryptoEventReasonSerializerTest.java
@@ -1,0 +1,52 @@
+package pl.zankowski.iextrading4j.client.mapper;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import pl.zankowski.iextrading4j.api.alternative.CryptoEventReason;
+
+import java.io.IOException;
+
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+
+public class CryptoEventReasonSerializerTest {
+
+    private CryptoEventReasonSerializer serializer;
+
+    @Before
+    public void setUp() {
+        serializer = new CryptoEventReasonSerializer();
+    }
+
+    @After
+    public void tearDown() {
+        serializer = null;
+    }
+
+    @Test
+    public void shouldWriteNullIfValueIsNull() throws IOException {
+        final JsonGenerator jsonGeneratorMock = mock(JsonGenerator.class);
+        final SerializerProvider serializerProviderMock = mock(SerializerProvider.class);
+        final CryptoEventReason input = null;
+
+        serializer.serialize(input, jsonGeneratorMock, serializerProviderMock);
+
+        verify(jsonGeneratorMock).writeNull();
+    }
+
+    @Test
+    public void shouldWriteStringFromValue() throws IOException {
+        final JsonGenerator jsonGeneratorMock = mock(JsonGenerator.class);
+        final SerializerProvider serializerProviderMock = mock(SerializerProvider.class);
+        final CryptoEventReason input = CryptoEventReason.INITIAL;
+
+        serializer.serialize(input, jsonGeneratorMock, serializerProviderMock);
+
+        verify(jsonGeneratorMock).writeString(eq("initial"));
+    }
+
+}