+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/mapper/CryptoSideDeserializerTest.java
@@ -1,0 +1,54 @@
+package pl.zankowski.iextrading4j.client.mapper;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import pl.zankowski.iextrading4j.api.alternative.CryptoSide;
+
+import java.io.IOException;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+public class CryptoSideDeserializerTest {
+
+    private CryptoSideDeserializer deserializer;
+
+    @Before
+    public void setUp() {
+        deserializer = new CryptoSideDeserializer();
+    }
+
+    @After
+    public void tearDown() {
+        deserializer = null;
+    }
+
+    @Test
+    public void shouldReturnUnknownTypeIfValueIsNull() throws IOException {
+        final JsonParser parserMock = mock(JsonParser.class);
+        final DeserializationContext contextMock = mock(DeserializationContext.class);
+
+        when(parserMock.getValueAsString()).thenReturn(null);
+
+        final CryptoSide result = deserializer.deserialize(parserMock, contextMock);
+
+        assertThat(result).isNull();
+    }
+
+    @Test
+    public void shouldCreateEnumBasedOnValue() throws IOException {
+        final JsonParser parserMock = mock(JsonParser.class);
+        final DeserializationContext contextMock = mock(DeserializationContext.class);
+
+        when(parserMock.getValueAsString()).thenReturn("bid");
+
+        final CryptoSide result = deserializer.deserialize(parserMock, contextMock);
+
+        assertThat(result).isEqualTo(CryptoSide.BID);
+    }
+
+}