+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/sse/request/alternative/CryptoQuoteSseRequestBuilderTest.java
@@ -1,28 +1,27 @@
 package pl.zankowski.iextrading4j.client.sse.request.alternative;
 
 import org.junit.Test;
-import pl.zankowski.iextrading4j.api.alternative.Sentiment;
+import pl.zankowski.iextrading4j.api.stocks.Quote;
 import pl.zankowski.iextrading4j.client.sse.manager.SseRequest;
 
 import javax.ws.rs.core.GenericType;
-
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
-public class SentimentSseRequestBuilderTest {
+public class CryptoQuoteSseRequestBuilderTest {
 
     @Test
     public void shouldSuccessfullyCreateRequest() {
         final String symbol = "IBM";
 
-        final SseRequest<List<Sentiment>> request = new SentimentSseRequestBuilder()
+        final SseRequest<List<Quote>> request = new CryptoQuoteSseRequestBuilder()
                 .withSymbol(symbol)
                 .build();
 
-        assertThat(request.getPath()).isEqualTo("/sentiment");
-        assertThat(request.getResponseType()).isEqualTo(new GenericType<List<Sentiment>>() {});
+        assertThat(request.getPath()).isEqualTo("/cryptoQuotes");
+        assertThat(request.getResponseType()).isEqualTo(new GenericType<List<Quote>>() {});
         assertThat(request.getPathParams()).isEmpty();
         assertThat(request.getQueryParams()).contains(entry("nosnapshot", "false"), entry("symbols", symbol));
     }